requirements.txt
Django==4.1.3
django-cpf==0.1.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
django-environ==0.9.0